Add form for adding books
