+++ src/components/books/card.js
@@ -1,0 +1,13 @@
+import BookDetails from './details';
+
+const BookCard = (item) => {
+  const { book } = item;
+
+  return (
+    <div>
+      <BookDetails book={book} />
+    </div>
+  );
+};
+
+export default BookCard;
