+++ src/components/books/index.js
@@ -1,47 +1,11 @@
-import BookDetails from './details';
+import BookList from './booklist';
+import AddingForm from './form';
 
-const books = [
-  {
-    id: 0,
-    author: 'Suzanne Collins',
-    genre: 'Action',
-    title: 'The Hunger Games',
-  },
-  {
-    id: 1,
-    author: 'Julie Carwood',
-    genre: 'Action',
-    title: 'Shadow Dance',
-  },
-  {
-    id: 2,
-    author: 'Jeffery Deaver',
-    genre: 'Thriller',
-    title: 'Shallow Graves',
-  },
-];
-
-const BookCard = (item) => {
-  const { book } = item;
-
-  return (
-    <div>
-      <BookDetails book={book} />
-    </div>
-  );
-};
-
-const BookList = () => (
+const Home = () => (
   <div>
-    <ul>
-      {books.map((book) => (
-        <li key={book.id}>
-          {' '}
-          <BookCard book={book} />
-        </li>
-      ))}
-    </ul>
+    <BookList />
+    <AddingForm />
   </div>
 );
 
-export default BookList;
+export default Home;
